requirements.txt
Flask==3.0.3
gunicorn==20.1.0 
python-dotenv==1.1.1